Creating hook for edit columns control. #136
