+++ explorer/app/config/model.ts
@@ -72,7 +72,8 @@ export interface SummaryConfig {
 }
 
 export interface ColumnConfig<
-  T,
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any -- This config model is part of a generic array
+  T = any,
   // eslint-disable-next-line @typescript-eslint/no-explicit-any -- This config model is part of a generic array
   C extends keyof JSX.IntrinsicElements | JSXElementConstructor<any> = any
 > {
