+++ explorer/site-config/hca-dcp/dev/projectViewModelBuilder.ts
@@ -223,17 +223,24 @@ export const projectsToCellCountColumn = (
   };
 };
 
-export const projectsToDevStageColumn = (
-  project: ProjectsResponse
+/**
+ * Build props for the Development stage Text component from the given projects response.
+ * @param projectsResponse - Response model return from projects API.
+ * @returns model to be used as props for the development stage table column.
+ */
+export const buildDevStage = (
+  projectsResponse: ProjectsResponse
 ): React.ComponentProps<typeof C.Text> => {
-  if (!project.donorOrganisms) {
+  if (!projectsResponse.donorOrganisms) {
     return {
       children: "",
     };
   }
   return {
     children: concatStrings(
-      project.donorOrganisms.flatMap((orgnanism) => orgnanism.developmentStage)
+      projectsResponse.donorOrganisms.flatMap(
+        (orgnanism) => orgnanism.developmentStage
+      )
     ),
     customColor: "ink",
     variant: "text-body-400",
